# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Website Live Chat Agent API
  version: 0.1.0
  description: |
    基于 LangGraph + Milvus + 插件化模型提供商的智能客服 Agent API。
    
    **核心特性**:
    - ✅ OpenAI Chat Completion API 完全兼容
    - ✅ 插件化模型提供商架构（支持OpenAI、DeepSeek、硅基流动等）
    - ✅ 模型配置分离（LLM和Embedding可独立选择不同提供商）
    - ✅ 混合模型组合支持（如DeepSeek LLM + OpenAI Embedding）
    - ✅ RAG 知识库检索（从 Milvus 检索网站内容）
    - ✅ 流式响应支持（SSE）
    - ✅ 会话状态持久化（LangGraph Checkpointer）
    
    **支持的模型提供商**:
    - **OpenAI**: gpt-4o-mini, gpt-4o, text-embedding-3-small
    - **DeepSeek**: deepseek-chat, deepseek-coder, deepseek-embedding
    - **硅基流动**: Qwen/Qwen2.5系列, BAAI/bge系列
    
    **WordPress 集成**:
    WordPress 插件只需修改 Base URL 为 `http://your-server:8000/v1` 即可使用。
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: http://{host}:8000
    description: 自定义部署
    variables:
      host:
        default: your-server
        description: 服务器地址
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/chat/completions:
    post:
      summary: 创建对话完成（OpenAI 兼容）
      description: |
        完全兼容 OpenAI Chat Completion API 格式。
        
        **内部流程**:
        1. LangGraph Router 判断是否需要检索知识库
        2. 如需检索，从 Milvus 检索 Top-3 相关文档
        3. 调用 DeepSeek LLM 生成响应
        4. 返回 OpenAI 格式的响应
        
        **支持流式响应**: 设置 `stream: true` 启用 SSE
      operationId: chatCompletions
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              simple_question:
                summary: 简单问答
                value:
                  model: deepseek-chat
                  messages:
                    - role: user
                      content: 你们的退货政策是什么？
                  stream: false
              
              multi_turn:
                summary: 多轮对话
                value:
                  model: deepseek-chat
                  messages:
                    - role: user
                      content: iPhone 15 有哪些颜色？
                    - role: assistant
                      content: iPhone 15 提供5种颜色：黑色、白色、蓝色、粉色、黄色。
                    - role: user
                      content: 那价格呢？
                  stream: false
              
              streaming:
                summary: 流式响应
                value:
                  model: deepseek-chat
                  messages:
                    - role: user
                      content: 介绍一下你们的产品
                  stream: true
                  temperature: 0.7

      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
              example:
                id: chatcmpl-abc123
                object: chat.completion
                created: 1699999999
                model: deepseek-chat
                choices:
                  - index: 0
                    message:
                      role: assistant
                      content: 我们的退货政策：收到商品后30天内可申请退货...
                    finish_reason: stop
                usage:
                  prompt_tokens: 15
                  completion_tokens: 50
                  total_tokens: 65
            
            text/event-stream:
              schema:
                type: string
                description: SSE 流式响应（stream=true 时）
              example: |
                data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1699999999,"model":"deepseek-chat","choices":[{"index":0,"delta":{"role":"assistant"},"finish_reason":null}]}
                
                data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1699999999,"model":"deepseek-chat","choices":[{"index":0,"delta":{"content":"我"},"finish_reason":null}]}
                
                data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1699999999,"model":"deepseek-chat","choices":[{"index":0,"delta":{"content":"们"},"finish_reason":null}]}
                
                data: [DONE]

        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== 知识库管理 ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API Key 认证，格式: `Bearer your-api-key`
        
        示例: `Authorization: Bearer sk-agent-12345abcde`
  schemas:
...
    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant]
          description: 消息角色
        content:
          type: string
          description: 消息内容
    ChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          default: deepseek-chat
          description: 模型名称（兼容性参数，实际使用配置中的 LLM）
          example: deepseek-chat
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          description: 对话历史
        stream:
          type: boolean
          default: false
          description: 是否启用流式响应（SSE）
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
          description: 温度参数（控制随机性）
        max_tokens:
          type: integer
          minimum: 1
          description: 最大生成 Token 数（可选）
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1.0
          description: 核采样参数
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: 响应ID
          example: chatcmpl-abc123
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          description: Unix 时间戳
          example: 1699999999
        model:
          type: string
          example: deepseek-chat
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionChoice'
        usage:
          $ref: '#/components/schemas/ChatCompletionUsage'
    ChatCompletionChoice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/ChatMessage'
        finish_reason:
          type: string
          enum: [stop, length, tool_calls]
          nullable: true
    ChatCompletionUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer

    # ===== 知识库模型 =====
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
  responses:
    Unauthorized:
      description: API Key 无效或缺失
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: Invalid API key
              type: invalid_request_error
              code: invalid_api_key
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: 'Validation error: messages is required'
              type: invalid_request_error
              code: validation_error
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: Internal server error
              type: server_error
              code: internal_error